requirements.txt
apache-airflow==2.8.1
snowflake-connector-python[pandas]
snowflake-sqlalchemy
dbt-snowflake
pandas
anthropic
streamlit
plotly
python-dotenv
kaggle
pinecone-client
